Add drivers directory and move files
