+++ andino_firmware/include/andino/drivers/encoder.h
