+++ andino_firmware/platformio.ini
@@ -36,8 +36,8 @@ platform = native
 test_framework = googletest
 test_build_src = true
 build_src_filter =
-  +<encoder.cpp>
-  +<motor.cpp>
+  +<drivers/encoder.cpp>
+  +<drivers/motor.cpp>
   +<pid.cpp>
   +<shell.cpp>
