+++ andino_firmware/src/app.cpp
@@ -73,10 +73,10 @@
 #include "commands.h"
 #include "constants.h"
 #include "andino/bsp/digital_out_arduino.h"
-#include "encoder.h"
+#include "andino/drivers/encoder.h"
 #include "hw.h"
 #include "andino/bsp/interrupt_in_arduino.h"
-#include "motor.h"
+#include "andino/drivers/motor.h"
 #include "pid.h"
 #include "andino/bsp/pwm_out_arduino.h"
 #include "andino/bsp/serial_stream_arduino.h"
