+++ andino_firmware/src/app.h
@@ -32,9 +32,9 @@
 #include <Adafruit_BNO055.h>
 
 #include "andino/bsp/digital_out_arduino.h"
-#include "encoder.h"
+#include "andino/drivers/encoder.h"
 #include "andino/bsp/interrupt_in_arduino.h"
-#include "motor.h"
+#include "andino/drivers/motor.h"
 #include "pid.h"
 #include "andino/bsp/pwm_out_arduino.h"
 #include "andino/bsp/serial_stream_arduino.h"
