+++ andino_firmware/test/desktop/test_motor/motor_test.cpp
@@ -27,7 +27,7 @@
 // CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 // OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 // OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-#include "motor.h"
+#include "andino/drivers/motor.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
